requirements.txt
streamlit
google-cloud-bigquery
google-generativeai
python-dotenv